requirements.txt
fastapi==0.111.1
uvicorn==0.25.0
pydantic==1.10.12
pymongo[srv]==4.7.1
python-dotenv==1.0.0
